Put the preview logic where this codebase keeps domain logic
entity/import_preview.py was the only file of its kind: no app here has a
module named after a feature. Where domain logic lives is entry/services.py,
as a service class. Both previews now live there -- EntityImportPreviewService
and EntryImportPreviewService -- and entry/tasks.py, which had grown by three
hundred lines of preview, is back to holding the task shells it holds for
everything else.

No behaviour changes.
